backend/requirements.txt
Flask
Flask-Cors
openai
python-dotenv
